rename test-common[-browser].js to test/common[-browser].js
