+++ test/common-browser.js
