+++ test/open-test.js
@@ -114,7 +114,7 @@ module.exports.tearDown = function (test, testCommon) {
 }
 
 module.exports.all = function (leveldown, test, testCommon) {
-  testCommon = testCommon || require('../test-common')
+  testCommon = testCommon || require('./common')
   module.exports.setUp(test, testCommon)
   module.exports.args(leveldown, test, testCommon)
   module.exports.open(leveldown, test, testCommon)
